Move POST /users/ to new api/users.js file #Fruehjahrsputz
