+++ server/package.json
@@ -19,7 +19,7 @@
     "angularjs-color-picker": "^3.3.0",
     "applicationinsights": "^1.4.1",
     "axios": "^1.5.0",
-    "bcryptjs": "^2.4.3",
+    "bcrypt": "5.x",
     "body-parser": "^1.17.2",
     "brute-knex": "^4.0.0",
     "colornames": "latest",
